+++ metaflow/plugins/kfp/tests/test_argo_support.py
@@ -1,28 +1,10 @@
 import os
 from tempfile import TemporaryDirectory
-from typing import List
 
 import pytest
 import subprocess_tee
 
-from metaflow import R
-
-
-def _python():
-    if R.use_r():
-        return "python3"
-    else:
-        return "python"
-
-
-def obtain_flow_file_paths(flow_dir_path: str) -> List[str]:
-    file_paths: List[str] = [
-        file_name
-        for file_name in os.listdir(flow_dir_path)
-        if os.path.isfile(os.path.join(flow_dir_path, file_name))
-        and not file_name.startswith(".")
-    ]
-    return file_paths
+from . import obtain_flow_file_paths
 
 
 @pytest.mark.parametrize("flow_file_path", obtain_flow_file_paths("flows"))
